track-time: move validation out of service and into TS for creating time tracking
