--- backend/src/time-tracks/domain/commands/create-time-track.command.ts
@@ -1,11 +1,0 @@
-export type CreateTimeTrackCommand = {
-  noteId: number;
-  date: Date;
-  startTime: string;
-  durationMinutes: number;
-  note?: string;
-  user: {
-    id: string;
-    username: string;
-  }
-}
