+++ backend/src/time-tracks/domain/transaction-scripts/create-time-track-TS/create-time-track.command.ts
@@ -1,8 +1,11 @@
 export type CreateTimeTrackCommand = {  
   noteId: number;
-  userId: string;
   date: Date;
   startTime: string;
   durationMinutes: number;
   note?: string;
+  user: {
+    id: string;
+    username: string;
+  }
 }
